Implement Error Parsing

+++ api/src/main/java/com/intel/i40eaqdebug/api/logs/LogEntry.java
@@ -4,7 +4,7 @@ public interface LogEntry {
 
     int getStartLine();
     long getTimeStampMillis();
-    byte getErr();
+    int getErr();
     short getFlags();
     short getOpCode();
     short getRetVal();
